speeding up produceexp_chance

+++ src/kernel/config.h
@@ -261,7 +261,7 @@ extern "C" {
         unsigned int data_turn;
         struct param *parameters;
         void *vm_state;
-        float producexpchance;
+        double producexpchance;
         int cookie;
         int data_version; /* TODO: eliminate in favor of gamedata.version */
         struct _dictionary_ *inifile;
